chore: Update database connection and add dotenv support
This commit updates the `Connect` function in `database.go` to include support for loading environment variables from a `.env` file using the `godotenv` package. It also retrieves the MongoDB URI from the `MONGOURI` environment variable. This change improves the flexibility and security of the database connection setup.

Additionally, the commit updates the `go.mod` and `go.sum` files to include the indirect dependency on the `github.com/joho/godotenv` package.

Lastly, the commit removes unnecessary commented code and adds a `.gitignore` file to ignore certain files and directories.

```